+++ scripts/dropbox-android.py
@@ -1,0 +1,9 @@
+# script for ease of use and ease of testing on android
+
+/post format:
+- if individual file, upload with filename as text input minus extension
+- if directory of files, upload each with the directory filename as text input
+
+/message format:
+- if individual file, check file for message specific metadata
+- if directory, check for message specific json file