+++ scripts/upload-to-dropbox.py
@@ -1,3 +1,6 @@
+# TODO: copy this in a way that uses the download method in place of the upload method, so reverse it
+
+
 # https://github.com/dropbox/dropbox-sdk-python/blob/main/example/updown.py
 
 """Upload the contents of your OnlySnarf Uploads folder to Dropbox.